requirements.txt
python-telegram-bot==20.3
scikit-learn==1.3.2
httpx==0.24.1
